+++ src/Services/InvoiceService.js
@@ -202,6 +202,15 @@ class InvoiceService
             throw new Error('Invoice does not exist.');
         }
 
+        // to double check and see if commodity is attached
+        const checkLine = await InvoiceLine.query().findById(lineGuid);
+
+        // if attached throw error
+        if (checkLine.itemId == 1 && checkLine.commodityGuid != null)
+        {
+            throw new Error('Cannot delete line.');
+        }
+
         // returning updated bill
         const newLine = await InvoiceLine.query().deleteById(lineGuid).returning('*');
